Poppy Voice Assistant Version 0.1.1
This is the first public version of my Python voice assistant Poppy! She's still got a long way to go, and I've still got a lot to add to her, but as is she is a functioning product and should be relatively easy to build upon if you are comfortable working with Python!